Modules/Devices/src/Devices.Application/Identities/Commands/StartDeletionProcessAsUser/Handler.cs
using Backbone.BuildingBlocks.Application.Abstractions.Exceptions;
using Backbone.BuildingBlocks.Application.Abstractions.Infrastructure.UserContext;
using Backbone.DevelopmentKit.Identity.Entities;
using Backbone.Modules.Devices.Application.Infrastructure.Persistence.Repository;
using MediatR;

namespace Backbone.Modules.Devices.Application.Identities.Commands.StartDeletionProcessAsUser;

public class Handler : IRequestHandler<StartDeletionProcessAsUserCommand, StartDeletionProcessAsUserResponse>
{
    private readonly IIdentitiesRepository _identitiesRepository;
    private readonly IUserContext _userContext;

    public Handler(IIdentitiesRepository identitiesRepository, IUserContext userContext)
    {
        _identitiesRepository = identitiesRepository;
        _userContext = userContext;
    }

    public async Task<StartDeletionProcessAsUserResponse> Handle(StartDeletionProcessAsUserCommand request, CancellationToken cancellationToken)
    {
        var identity = await _identitiesRepository.FindByAddress(_userContext.GetAddress(), cancellationToken, true) ?? throw new NotFoundException(nameof(Identity));

        var deletionProcess = identity.StartDeletionProcess(_userContext.GetDeviceId());

        await _identitiesRepository.Update(identity, cancellationToken);

        return new StartDeletionProcessAsUserResponse(deletionProcess);
    }
}
